Move generic image stuff to core, etc
