+++ proseco/core.py
@@ -8,7 +8,7 @@ import yaml
 from scipy.interpolate import interp1d
 from astropy.table import Table, Column
 
-from chandra_aca.transform import yagzag_to_pixels
+from chandra_aca.transform import yagzag_to_pixels, count_rate_to_mag
 from Ska.quatutil import radec2yagzag
 from agasc import get_agasc_cone
 from Quaternion import Quat
@@ -22,9 +22,9 @@ def to_python(val):
     return val
 
 
-class AcqTable(Table):
+class ACACatalogTable(Table):
     def __init__(self, *args, print_log=False, **kwargs):
-        super(AcqTable, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self.log_info = {}
         self.log_info['events'] = []
         self.log_info['time0'] = time.time()
@@ -72,7 +72,7 @@ class AcqTable(Table):
             if name in names:
                 self[name].format = self._default_formats.pop(name)
 
-        return super(AcqTable, self[names])._base_repr_(*args, **kwargs)
+        return super(ACACatalogTable, self[names])._base_repr_(*args, **kwargs)
 
     def to_yaml(self, rootdir=None):
         """
@@ -115,7 +115,7 @@ class AcqTable(Table):
         return acqs
 
     def to_struct(self):
-        """Turn table into a dict structure with keys:
+        """Turn table ``tbl`` into a dict structure with keys:
 
         - names: column names in order
         - dtype: column dtypes as strings
@@ -137,7 +137,7 @@ class AcqTable(Table):
                     val = Table(val)
 
                 if isinstance(val, Table):
-                    val = AcqTable.to_struct(val)
+                    val = ACACatalogTable.to_struct(val)
 
                 elif isinstance(val, dict):
                     new_val = {}
@@ -275,3 +275,72 @@ def get_stars(att, date=None, radius=1.2, logger=None):
     logger('Finished star processing', level=1)
 
     return stars
+
+
+def bin2x2(arr):
+    """Bin 2-d ``arr`` in 2x2 blocks.  Requires that ``arr`` has even shape sizes"""
+    shape = (arr.shape[0] // 2, 2, arr.shape[1] // 2, 2)
+    return arr.reshape(shape).sum(-1).sum(1)
+
+
+RC_6x6 = np.array([10, 11, 12, 13,
+                   17, 18, 19, 20, 21, 22,
+                   25, 26, 27, 28, 29, 30,
+                   33, 34, 35, 36, 37, 38,
+                   41, 42, 43, 44, 45, 46,
+                   50, 51, 52, 53],
+                  dtype=np.int64)
+
+ROW_6x6 = np.array([1, 1, 1, 1,
+                    2, 2, 2, 2, 2, 2,
+                    3, 3, 3, 3, 3, 3,
+                    4, 4, 4, 4, 4, 4,
+                    5, 5, 5, 5, 5, 5,
+                    6, 6, 6, 6],
+                   dtype=np.float64) + 0.5
+
+COL_6x6 = np.array([2, 3, 4, 5,
+                    1, 2, 3, 4, 5, 6,
+                    1, 2, 3, 4, 5, 6,
+                    1, 2, 3, 4, 5, 6,
+                    1, 2, 3, 4, 5, 6,
+                    2, 3, 4, 5],
+                   dtype=np.float64) + 0.5
+
+
+def get_image_props(ccd_img, c_row, c_col, bgd=None):
+    """
+    Do a pseudo-read and compute the background-subtracted magnitude
+    of the image.  Returns the 8x8 image and mag.
+
+    :param ccd_image: full-frame CCD image (e.g. dark current, with or without stars).
+    :param c_row: int row at center (readout row-4 to row+4)
+    :param c_col: int col at center
+    :param bgd: background to subtract at each pixel.  If None then compute flight bgd.
+
+    :returns: 8x8 image (ndarray), image mag
+    """
+    img = ccd_img[c_row - 4:c_row + 4, c_col - 4:c_col + 4]
+
+    if bgd is None:
+        raise NotImplementedError('no calc_flight_background here')
+        # bgd = calc_flight_background(img)  # currently not defined
+
+    img = img - bgd
+    img_6x6 = img.flatten()[RC_6x6]
+    img_sum = np.sum(img_6x6)
+    r = np.sum(img_6x6 * ROW_6x6) / img_sum
+    c = np.sum(img_6x6 * COL_6x6) / img_sum
+    row = r + c_row - 4
+    col = c + c_col - 4
+    mag = count_rate_to_mag(img_sum)
+
+    return img, img_sum, mag, row, col
+
+
+def pea_reject_image(img):
+    """
+    Check if PEA would reject image (too narrow, too peaky, etc)
+    """
+    # To be implemented
+    return False
